Спринт 23 домашка

+++ app/src/main/java/com/practicum/playlistmaker/search/data/repository/RepositoryNetworkRetrofitImpl.kt
@@ -28,7 +28,7 @@ class RepositoryNetworkRetrofitImpl(
                             it.trackId,
                             it.trackName,
                             it.artistName,
-                            it.trackTimeMillis,
+                            it.trackTimeMillis.toLong(),
                             it.artworkUrl100,
                             it.collectionName,
                             it.releaseDate,
